backend/requirements.txt
Authlib==1.6.5
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.6.0
Werkzeug==3.1.3
zipp==3.23.0
psycopg2-binary==2.9.10